User Profile

+++ study_budy/base/views.py
@@ -7,6 +7,8 @@ from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate,login,logout
 from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+from .models import Message
 # rooms = [
 #     {'id':1,'name':'Lets learn Python!'},
 #     {'id':2,'name':'Design with me'},
@@ -18,13 +20,14 @@ from django.contrib.auth.decorators import login_required
 # If user is authenticated it creates a session for the user
 # If authentication fails it shows error message using django messages framework
 def loginPage(request):
+    page = 'login'
     if request.user.is_authenticated:
         return redirect('home')
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
         try:
-            user = User.objects.get(username=username)
+            user = User.objects.get(username=username.lower())
         except:
             messages.error(request, 'User does not exist')
         user = authenticate(request,username=username,password=password)
@@ -33,12 +36,29 @@ def loginPage(request):
             return redirect('home')
         else:
             messages.error(request,'Username OR password does not exist')
-    context = {}
+    context = {'page':page}
     return render(request,"base/login_register.html",context)
 
 def logoutUser(request):
     logout(request)
     return redirect('home')
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request,user)
+            return redirect('home')
+        else: 
+            messages.error(request,'An error occured during registration')
+
+    return render(request,'base/login_register.html',{'form':form})
+
+
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
     rooms = Room.objects.filter(
@@ -47,13 +67,25 @@ def home(request):
         Q(description__icontains=q))
     topics = Topic.objects.all()
     room_count = rooms.count()
+    room_messages = Message.objects.filter(Q(room__topic__name__icontains=q))
 
-    context = {'rooms':rooms,'topics': topics , 'room_count': room_count}
+    context = {'rooms':rooms,'topics': topics , 'room_count': room_count, 'room_messages': room_messages}
     return render(request,'base/home.html',context)
 
 def room(request,pk):
     room = Room.objects.get(id=pk)
-    context = {'room':room}
+    room_m = room.message_set.all().order_by('-created')
+    participants = room.participants.all() #get all participants in the room
+    if request.method == 'POST':
+        message = Message.objects.create(
+            user=request.user,
+            room=room,
+            body=request.POST.get('body')
+        )
+        room.participants.add(request.user)  #add user to participants when they post a message
+        return redirect('room',pk=room.id)
+        
+    context = {'room':room, 'room_m':room_m , 'participants':participants}
     return render(request,'base/room.html',context)
 
 #This decorator restricts access to createRoom view only to authenticated users and sends them to login page if not authenticated
@@ -87,7 +119,25 @@ def updateRoom(request,pk):
 @login_required(login_url='login')
 def deleteRoom(request,pk):
     room  = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+@login_required(login_url='login')
+def deleteMessage(request,pk):
+    message  = Message.objects.get(id=pk)
+    room_id = message.room.id
+    if request.user != message.user:
+        return HttpResponse('You are not allowed here!!')
+    if request.method == 'POST':
+        message.delete()
+        return redirect('room',pk=room_id)
+    return render(request,'base/delete.html',{'obj':message})
+
+def  userProfile(request,pk):
+    user = User.objects.get(id=pk)
+    context = {'user': user}
+    return render(request,'base/profile.html',context)
